Move EBS snapshot copt to S3 test into e2e_cli.py
